refactor speculative tests
