--- mlx_vlm/tests/test_drafter_kind_resolution.py
@@ -1,92 +1,0 @@
-"""Auto-detect drafter ``kind`` from the drafter's HF ``model_type``.
-
-Regression for the trap where ``--draft-model`` points at a
-``gemma4_assistant`` checkpoint but ``--draft-kind`` is left at the default
-``"dflash"``: the DFlash round-loop would then call ``draft_block`` on a
-drafter that needs the MTP round-loop, raising an opaque
-``set_shared_kv() must be called`` error from inside generate.py.
-"""
-
-import json
-from pathlib import Path
-
-import pytest
-
-from mlx_vlm.speculative.drafters import (
-    DEFAULT_DRAFTER_KIND,
-    DRAFTER_KIND_BY_MODEL_TYPE,
-    KNOWN_DRAFTER_KINDS,
-    resolve_drafter_kind,
-)
-
-
-def _make_drafter_dir(tmp_path: Path, model_type: str | None) -> Path:
-    d = tmp_path / "drafter"
-    d.mkdir()
-    cfg = {} if model_type is None else {"model_type": model_type}
-    (d / "config.json").write_text(json.dumps(cfg))
-    return d
-
-
-def test_gemma4_assistant_overrides_dflash_to_mtp(tmp_path, caplog):
-    path = _make_drafter_dir(tmp_path, "gemma4_assistant")
-    with caplog.at_level("WARNING"):
-        resolved = resolve_drafter_kind(path, "dflash")
-    assert resolved == "mtp"
-    assert any("requires --draft-kind='mtp'" in r.getMessage() for r in caplog.records)
-
-
-def test_explicit_mtp_kept_for_gemma4_assistant(tmp_path, caplog):
-    path = _make_drafter_dir(tmp_path, "gemma4_assistant")
-    with caplog.at_level("WARNING"):
-        resolved = resolve_drafter_kind(path, "mtp")
-    assert resolved == "mtp"
-    assert caplog.records == []  # no warning when kinds already align
-
-
-def test_unknown_model_type_keeps_caller_kind(tmp_path):
-    path = _make_drafter_dir(tmp_path, "qwen3_dflash")
-    assert resolve_drafter_kind(path, "dflash") == "dflash"
-
-
-def test_missing_config_keeps_caller_kind(tmp_path):
-    d = tmp_path / "drafter"
-    d.mkdir()
-    # No config.json at all — resolver should fall back gracefully.
-    assert resolve_drafter_kind(d, "dflash") == "dflash"
-
-
-def test_malformed_config_keeps_caller_kind(tmp_path):
-    d = tmp_path / "drafter"
-    d.mkdir()
-    (d / "config.json").write_text("not valid json {")
-    assert resolve_drafter_kind(d, "dflash") == "dflash"
-
-
-def test_kind_table_only_uses_known_kinds():
-    # Every override must be a recognized round-loop kind.
-    for mt, kind in DRAFTER_KIND_BY_MODEL_TYPE.items():
-        assert kind in KNOWN_DRAFTER_KINDS, f"{mt} maps to unknown kind {kind}"
-
-
-@pytest.mark.parametrize("model_type", [None])
-def test_resolver_handles_no_model_type_field(tmp_path, model_type):
-    path = _make_drafter_dir(tmp_path, model_type)
-    assert resolve_drafter_kind(path, "dflash") == "dflash"
-
-
-def test_kind_none_autodetects_mtp_for_gemma4_assistant(tmp_path):
-    path = _make_drafter_dir(tmp_path, "gemma4_assistant")
-    assert resolve_drafter_kind(path, None) == "mtp"
-    assert resolve_drafter_kind(path) == "mtp"  # default arg
-
-
-def test_kind_none_falls_back_to_default_for_unknown_model_type(tmp_path):
-    path = _make_drafter_dir(tmp_path, "qwen3_dflash")
-    assert resolve_drafter_kind(path, None) == DEFAULT_DRAFTER_KIND
-
-
-def test_kind_none_falls_back_to_default_for_missing_config(tmp_path):
-    d = tmp_path / "drafter"
-    d.mkdir()
-    assert resolve_drafter_kind(d, None) == DEFAULT_DRAFTER_KIND
