--- mlx_vlm/tests/test_gemma4_assistant_masks.py
@@ -1,90 +1,0 @@
-import mlx.core as mx
-
-from mlx_vlm.speculative.drafters.gemma4_assistant.masks import (
-    make_drafter_masks,
-    normalize_batched_shared_kv_states,
-)
-
-
-def test_mtp_drafter_masks_support_batched_offsets():
-    kv = (mx.zeros((2, 1, 8, 4)), mx.zeros((2, 1, 8, 4)))
-
-    masks = make_drafter_masks(
-        {"sliding_attention": kv, "full_attention": kv},
-        query_len=1,
-        query_offset=mx.array([5, 8]),
-        sliding_window=4,
-    )
-
-    assert masks["sliding_attention"].shape == (2, 1, 1, 8)
-    assert masks["full_attention"].shape == (2, 1, 1, 8)
-    full = masks["full_attention"].tolist()
-    assert full[0][0][0][5] == -float("inf")
-    assert full[1][0][0][7] == 0.0
-
-
-def test_normalize_batched_shared_kv_states_repacks_left_padded_rows():
-    keys = mx.array(
-        [
-            [[[0], [0], [0], [10], [11], [12], [13], [14]]],
-            [[[20], [21], [22], [23], [24], [25], [26], [27]]],
-        ],
-        dtype=mx.float32,
-    )
-    values = keys + 100
-
-    normalized = normalize_batched_shared_kv_states(
-        {"full_attention": (keys, values)},
-        kv_valid_len=mx.array([5, 8]),
-        left_padding=mx.array([3, 0]),
-    )
-
-    norm_keys, norm_values = normalized["full_attention"]
-    assert norm_keys[:, 0, :, 0].tolist() == [
-        [10.0, 11.0, 12.0, 13.0, 14.0, 0.0, 0.0, 0.0],
-        [20.0, 21.0, 22.0, 23.0, 24.0, 25.0, 26.0, 27.0],
-    ]
-    assert norm_values[:, 0, :, 0].tolist() == [
-        [110.0, 111.0, 112.0, 113.0, 114.0, 0.0, 0.0, 0.0],
-        [120.0, 121.0, 122.0, 123.0, 124.0, 125.0, 126.0, 127.0],
-    ]
-
-
-def test_normalize_batched_shared_kv_states_drops_tail_zero_slack():
-    keys = mx.array(
-        [
-            [[[0], [0], [0], [10], [11], [12], [13], [14], [15], [16], [17], [0]]],
-            [[[20], [21], [22], [23], [24], [25], [26], [27], [28], [29], [30], [31]]],
-        ],
-        dtype=mx.float32,
-    )
-    values = keys + 100
-
-    normalized = normalize_batched_shared_kv_states(
-        {"sliding_attention": (keys, values)},
-        kv_valid_len=mx.array([8, 12]),
-        left_padding=mx.array([3, 0]),
-    )
-
-    norm_keys, norm_values = normalized["sliding_attention"]
-    assert norm_keys[:, 0, :, 0].tolist() == [
-        [10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 17.0, 0.0, 0.0, 0.0, 0.0],
-        [20.0, 21.0, 22.0, 23.0, 24.0, 25.0, 26.0, 27.0, 28.0, 29.0, 30.0, 31.0],
-    ]
-    assert norm_values[:, 0, :, 0].tolist() == [
-        [110.0, 111.0, 112.0, 113.0, 114.0, 115.0, 116.0, 117.0, 0.0, 0.0, 0.0, 0.0],
-        [
-            120.0,
-            121.0,
-            122.0,
-            123.0,
-            124.0,
-            125.0,
-            126.0,
-            127.0,
-            128.0,
-            129.0,
-            130.0,
-            131.0,
-        ],
-    ]
