--- mlx_vlm/tests/test_qwen3_5_language.py
@@ -1,114 +1,0 @@
-from types import SimpleNamespace
-from unittest.mock import patch
-
-import mlx.core as mx
-
-import mlx_vlm.models.qwen3_5.language as qwen_language
-from mlx_vlm.models.cache import ArraysCache
-
-
-def _make_conv_input(batch_size: int, layer_offset: int, length: int = 5) -> mx.array:
-    rows = []
-    for row in range(batch_size):
-        rows.append([[layer_offset * 100 + row * 10 + t] for t in range(length)])
-    return mx.array(rows, dtype=mx.float32)
-
-
-def _make_gdn_state(
-    batch_size: int, layer_offset: int, *, init_state: mx.array | None
-) -> tuple:
-    q = mx.full((batch_size, 3, 3, 4), layer_offset + 0.1, dtype=mx.float32)
-    k = mx.full((batch_size, 3, 3, 4), layer_offset + 0.2, dtype=mx.float32)
-    v = mx.full((batch_size, 3, 3, 5), layer_offset + 0.3, dtype=mx.float32)
-    a = mx.full((batch_size, 3, 3), layer_offset + 0.4, dtype=mx.float32)
-    b = mx.full((batch_size, 3, 3), layer_offset + 0.5, dtype=mx.float32)
-    A_log = mx.full((3,), layer_offset + 0.6, dtype=mx.float32)
-    dt_bias = mx.full((3,), layer_offset + 0.7, dtype=mx.float32)
-    conv_input = _make_conv_input(batch_size, layer_offset)
-    return (
-        q,
-        k,
-        v,
-        a,
-        b,
-        A_log,
-        dt_bias,
-        init_state,
-        None,
-        conv_input,
-        4,
-    )
-
-
-def test_qwen_rollback_speculative_cache_flattens_batch_per_layer():
-    batch_size = 2
-    accepted = mx.array([0, 1], dtype=mx.int32)
-    caches = [ArraysCache(size=2), ArraysCache(size=2)]
-    state0 = mx.full((batch_size, 3, 5, 4), 10.0, dtype=mx.float32)
-    state1 = mx.full((batch_size, 3, 5, 4), 20.0, dtype=mx.float32)
-    gdn_states = [
-        _make_gdn_state(batch_size, 0, init_state=state0),
-        _make_gdn_state(batch_size, 1, init_state=state1),
-    ]
-    captured = {}
-
-    def fake_gated_delta_update(
-        q, k, v, a, b, A_log, dt_bias, state, mask, use_kernel=True
-    ):
-        del k, v, a, b, use_kernel
-        captured["q_shape"] = q.shape
-        captured["A_log_shape"] = A_log.shape
-        captured["dt_bias_shape"] = dt_bias.shape
-        captured["mask"] = mask
-        row_ids = mx.arange(state.shape[0], dtype=mx.float32).reshape(-1, 1, 1, 1)
-        states_out = mx.broadcast_to(row_ids, state.shape)
-        return mx.zeros((q.shape[0], q.shape[1], 3, 5), dtype=mx.float32), states_out
-
-    with patch.object(
-        qwen_language, "gated_delta_update", side_effect=fake_gated_delta_update
-    ):
-        max_a = qwen_language.LanguageModel.rollback_speculative_cache(
-            SimpleNamespace(), caches, gdn_states, accepted, block_size=3
-        )
-
-    assert max_a == 1
-    assert captured["q_shape"] == (4, 2, 3, 4)
-    assert captured["A_log_shape"] == (4, 1, 3)
-    assert captured["dt_bias_shape"] == (4, 1, 3)
-    assert captured["mask"].tolist() == [
-        [True, False],
-        [True, True],
-        [True, False],
-        [True, True],
-    ]
-    assert caches[0][1][:, 0, 0, 0].tolist() == [0.0, 1.0]
-    assert caches[1][1][:, 0, 0, 0].tolist() == [2.0, 3.0]
-    assert caches[0][0][:, :, 0].tolist() == [[1.0, 2.0, 3.0], [12.0, 13.0, 14.0]]
-    assert caches[1][0][:, :, 0].tolist() == [
-        [101.0, 102.0, 103.0],
-        [112.0, 113.0, 114.0],
-    ]
-
-
-def test_qwen_rollback_speculative_cache_zero_inits_missing_state():
-    accepted = mx.array([1, 0], dtype=mx.int32)
-    caches = [ArraysCache(size=2)]
-    gdn_states = [_make_gdn_state(batch_size=2, layer_offset=0, init_state=None)]
-    captured = {}
-
-    def fake_gated_delta_update(
-        q, k, v, a, b, A_log, dt_bias, state, mask, use_kernel=True
-    ):
-        del q, k, v, a, b, A_log, dt_bias, mask, use_kernel
-        captured["state"] = state
-        return mx.zeros((state.shape[0], 2, 3, 5), dtype=mx.float32), state
-
-    with patch.object(
-        qwen_language, "gated_delta_update", side_effect=fake_gated_delta_update
-    ):
-        qwen_language.LanguageModel.rollback_speculative_cache(
-            SimpleNamespace(), caches, gdn_states, accepted, block_size=3
-        )
-
-    assert captured["state"].shape == (2, 3, 5, 4)
-    assert float(mx.sum(mx.abs(captured["state"])).item()) == 0.0
